requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-jose[cryptography]
authlib
pydantic
redis
arq
structlog
python-dotenv
email-validator
